restructure files in data_modules
