+++ pytorch_forecasting/data/data_module/__init__.py
@@ -1,17 +1,15 @@
 """Data Modules (D2 Layer) of pytorch-forecasting v2"""
 
-from pytorch_forecasting.data.data_module._base_data_module import (
+from pytorch_forecasting.data.data_module.base import (
     BaseTimeSeriesDataModule,
+    _BasePtDataModule,
 )
-from pytorch_forecasting.data.data_module._base_datamodule_pkg import _BasePtDataModule
-from pytorch_forecasting.data.data_module._encoder_decoder_data_module import (
+from pytorch_forecasting.data.data_module.encoder_decoder import (
+    EncoderDecoderDataModule_pkg,
     EncoderDecoderTimeSeriesDataModule,
 )
-from pytorch_forecasting.data.data_module._encoder_decoder_datamodule_pkg import (
-    EncoderDecoderDataModule_pkg,
-)
-from pytorch_forecasting.data.data_module._tslib_data_module import TslibDataModule
-from pytorch_forecasting.data.data_module._tslib_datamodule_pkg import (
+from pytorch_forecasting.data.data_module.tslib import (
+    TslibDataModule,
     TslibDataModule_pkg,
 )
