+++ pytorch_forecasting/data/data_module/base/__init__.py
@@ -1,0 +1,13 @@
+"""Base D2 datamodule components."""
+
+from pytorch_forecasting.data.data_module.base._data_module import (
+    NORMALIZER,
+    BaseTimeSeriesDataModule,
+)
+from pytorch_forecasting.data.data_module.base._datamodule_pkg import _BasePtDataModule
+
+__all__ = [
+    "NORMALIZER",
+    "BaseTimeSeriesDataModule",
+    "_BasePtDataModule",
+]
