+++ pytorch_forecasting/data/data_module/encoder_decoder/__init__.py
@@ -1,0 +1,13 @@
+"""Encoder-decoder D2 datamodule components."""
+
+from pytorch_forecasting.data.data_module.encoder_decoder._data_module import (
+    EncoderDecoderTimeSeriesDataModule,
+)
+from pytorch_forecasting.data.data_module.encoder_decoder._datamodule_pkg import (
+    EncoderDecoderDataModule_pkg,
+)
+
+__all__ = [
+    "EncoderDecoderTimeSeriesDataModule",
+    "EncoderDecoderDataModule_pkg",
+]
