+++ pytorch_forecasting/data/data_module/encoder_decoder/_data_module.py
@@ -13,11 +13,11 @@ from sklearn.preprocessing import RobustScaler, StandardScaler
 import torch
 from torch.utils.data import Dataset
 
-from pytorch_forecasting.data.data_module._base_data_module import (
+from pytorch_forecasting.data.data_module.base._data_module import (
     NORMALIZER,
     BaseTimeSeriesDataModule,
 )
-from pytorch_forecasting.data.data_module._encoder_decoder_dataset import (
+from pytorch_forecasting.data.data_module.encoder_decoder._dataset import (
     _ProcessedEncoderDecoderDataset,
 )
 from pytorch_forecasting.data.encoders import (
