+++ pytorch_forecasting/data/data_module/tslib/_datamodule_pkg.py
@@ -1,6 +1,6 @@
 """Package container for TslibDataModule."""
 
-from pytorch_forecasting.data.data_module._base_datamodule_pkg import _BasePtDataModule
+from pytorch_forecasting.data.data_module.base._datamodule_pkg import _BasePtDataModule
 
 
 class TslibDataModule_pkg(_BasePtDataModule):
